ai_module/requirements.txt
openai>=1.0.0
python-dotenv
psycopg2-binary
pgvector
